Split examples into own file
